Happy birthday to me, happy birthday to me... Well, it's not really
my birthday, but it will be in like 8 months and for such a huge celebration
I want to start preparing rather early.
The marshelling in Message is crap (not the good kind of crap either). I'm
not yet sure how we'll handle the transition from dcop but for now lets
at least make it sane api wise.